refactor(go-template): extract component-tree analysis into analysis/
Move the pure IR structural walks — `hasClientInteractivity`,
`hasEventsInTree`, child/nested-component discovery — out of the adapter
into `analysis/component-tree.ts` as free functions. These read no adapter
state and do no rendering, so they need no context. First step of the
Phase 4 decomposition. Output byte-identical (unit 551, conformance 786).
